Uploaded latest files
Uploaded the latest files with dates amended to file names for clarity.  The shield files with the F indicate the board for the Front of the Trinket, while the files with the B indicate the board for the Back of the Trinket.  The front board was also modified with a wider cutout for the micro USB port.